finished moving each area to its own file
ran into a circular import error witch i am still working on
