+++ court_yard.py
@@ -1,6 +1,6 @@
-from choices import get_choice3
+from redirect import c3
 
 
 def court_yard():
     print("you look around and see a well and a open gate")
-    get_choice3()
+    c3()
