+++ entrance_hall.py
@@ -1,0 +1,2 @@
+def entrance_hall():
+    print("you look around and see some stairs leading up to a higher floor")
