+++ game_start.py
@@ -1,0 +1,6 @@
+from name import name
+
+
+def game_start():
+    print("welcome to my text adventure")
+    print("welcome " + name + " to the castle jackson")
