+++ main_doors.py
@@ -1,0 +1,6 @@
+from redirect import c1
+
+
+def main_doors():
+    print("you head back to where you started")
+    c1()
