+++ open_gate.py
@@ -1,0 +1,2 @@
+def open_gate():
+    print("you see that there is a dense forest through the gate")
