+++ redirect.py
@@ -1,0 +1,13 @@
+from choices import get_choice1, get_choice2, get_choice3
+
+
+def c1():
+    get_choice1
+
+
+def c2():
+    get_choice2
+
+
+def c3():
+    get_choice3()
